Fixed a large number of warnings when compiled under C++23 with -Wall - C++20 deprecates implicit enum to float conversion and enum/enum arithmetic - Some variables that appeared to be used uninitialized were explicitly initialized to 0/NULL - A few member initializers were out of order, they were put in class order - some incorrect sprintf format specifiers were fixed - some unsigned/int comparisons were coerced to int/int - Switch [[fallthrough]] is used if case fall-through is desired. This uses a new   AGG_FALLTHROUGH symbol

+++ examples/mol_view.cpp
@@ -931,7 +931,7 @@ public:
         ls.width(1.5);
         label.text(mol.name());
         label.size(10.0);
-        label.start_point(10.0, start_height - 20.0);
+        label.start_point(10.0, (double)start_height - 20.0);
         ras.reset();
         ras.add_path(name);
         rs.color(agg::rgba(0,0,0));
